Swimming intermediate implementation
Functional intermediate swimming system. Working bobbing for idle player based on wave's altitude/frequency, implemented variable bobbing depending on submerged depth, nullifying it when the player hits the bottom of the sea.

Needs some small correction when on surface to make it more real and tweak the jumping force to get out of water.